# shill-guard/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShillGuard API
  description: |
    ShillGuard 水军智能检测系统接口文档。
    
    ## 认证说明
    登录成功后获取 `token`，后续所有需要认证的接口请在请求头添加：
    ```
    Authorization: Bearer <token>
    ```
    网关会自动解析 token 并注入 `X-User-Id`、`X-User-Role` 头，微服务内部通过这两个头识别用户身份。
    
    ## 角色权限
    | role 值 | 说明 |
    |---------|------|
    | 0 | 普通用户 |
    | 1 | 版主 |
    | 2 | 管理员 |
    
    ## 统一响应格式
    所有接口均返回统一格式：
    ```json
    {
      "code": 200,
      "message": "操作成功",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: ShillGuard Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境（API 网关）
  - url: http://localhost:3001/api
    description: 前端代理（开发调试）
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录后获取的 JWT Token
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: 操作成功
        data:
          description: 响应数据（类型随接口不同而变化）
    ResultVoid:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              example: null

    # 分页结构
    PageInfo:
      type: object
      properties:
        records:
          type: array
          description: 数据列表
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 128
        size:
          type: integer
          description: 每页大小
          example: 10
        current:
          type: integer
          description: 当前页码
          example: 1
        pages:
          type: integer
          description: 总页数
          example: 13

    # 登录请求
...
    CommentCreateDTO:
      type: object
      required:
        - postId
        - content
      properties:
        postId:
          type: integer
          format: int64
          description: 所属帖子ID（必填）
          example: 2001
        content:
          type: string
          description: 评论内容（必填）
          example: 写得真好，学到了！
        parentCommentId:
          type: integer
          format: int64
          nullable: true
          description: 父评论ID（为null则为一级评论，有值则为回复）
          example: null
        replyToUserId:
          type: integer
          format: int64
          nullable: true
          description: 被回复用户ID（回复评论时填写）
          example: null

    # 评论实体
    ContentComment:
      type: object
      properties:
        commentId:
          type: integer
          format: int64
          description: 评论ID
          example: 3001
        userId:
          type: integer
          format: int64
          description: 评论者用户ID
          example: 1001
        username:
          type: string
          description: 评论者用户名
          example: user01
        nickname:
          type: string
          description: 评论者昵称
          example: 小红
        avatarUrl:
          type: string
          nullable: true
          description: 评论者头像
          example: "http://localhost:9000/shill-guard/avatars/a.jpg"
        postId:
          type: integer
          format: int64
          description: 所属帖子ID
          example: 2001
        parentCommentId:
          type: integer
          format: int64
          nullable: true
          description: 父评论ID（null=一级评论）
          example: null
        replyToUserId:
          type: integer
          format: int64
          nullable: true
          description: 被回复用户ID
          example: null
        content:
          type: string
          description: 评论内容
          example: 写得真好，学到了！
        likeCount:
          type: integer
          description: 点赞数
          example: 5
        status:
          type: integer
          description: 状态（0=正常 1=删除）
          example: 0
        createdTime:
          type: string
          format: date-time
          description: 评论时间
          example: "2024-06-15T15:00:00"

    # 禁言记录实体
    AgentMuteRecord:
      type: object
      properties:
        muteId:
          type: integer
          format: int64
          description: 禁言记录ID
          example: 4001
        mutedUserId:
          type: integer
          format: int64
          description: 被禁言用户ID
          example: 1005
        reporterUserId:
          type: integer
          format: int64
          nullable: true
          description: 举报人用户ID（可为空）
          example: null
        reviewerUserId:
          type: integer
          format: int64
          nullable: true
          description: 审核人ID（人工操作时有值）
          example: 1001
        relatedCommentId:
          type: integer
          format: int64
          nullable: true
          description: 关联违规评论ID
          example: 3005
        muteReason:
          type: string
          description: 禁言原因
          example: 系统检测到该用户属于水军群组，协同刷评论
        muteDays:
          type: integer
          description: 禁言天数
          example: 7
        muteType:
          type: integer
          description: 禁言类型（0=Agent自动 1=人工）
          enum: [0, 1]
          example: 0
        status:
          type: integer
          description: 状态（1=禁言中 0=已解除）
          enum: [0, 1]
          example: 1
        createdTime:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-06-15T16:00:00"
        expireTime:
          type: string
          format: date-time
          nullable: true
          description: 解除时间
          example: "2024-06-22T16:00:00"

    # 错误响应
    ErrorResult:
      type: object
      properties:
        code:
          type: integer
          description: 错误码
          example: 401
        message:
          type: string
          description: 错误描述
          example: 未登录或Token已过期
        data:
          type: object
          nullable: true
          example: null

# ===================== 接口定义 =====================
paths:
...
  /api/content/comments/post/{postId}:
    get:
      tags: [评论]
      summary: 获取帖子的评论列表（分页）
      description: 获取指定帖子下的一级评论列表，按时间排序。无需登录。
      operationId: getCommentsByPost
      parameters:
        - name: postId
          in: path
          required: true
          description: 帖子ID
          schema:
            type: integer
            format: int64
            example: 2001
        - name: pageNum
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            default: 1
            example: 1
        - name: pageSize
          in: query
          required: false
          description: 每页数量
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageInfo'
                          - type: object
                            properties:
                              records:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ContentComment'
              example:
                code: 200
                message: 操作成功
                data:
                  records:
                    - commentId: 3001
                      userId: 1002
                      nickname: "小明"
                      content: "写得真好！"
                      likeCount: 5
                      createdTime: "2024-06-15T15:00:00"
                  total: 23
                  size: 20
                  current: 1
  /api/content/comments:
    post:
      tags: [评论]
      summary: 发表评论
      description: |
        发表评论或回复。
        - `parentCommentId` 为空时，发表一级评论。
        - `parentCommentId` 有值时，回复指定评论（二级评论）。
        需要登录。
      operationId: addComment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateDTO'
            examples:
              一级评论:
                value:
                  postId: 2001
                  content: "写得真好，学到了！"
              回复评论:
                value:
                  postId: 2001
                  content: "同意楼上！"
                  parentCommentId: 3001
                  replyToUserId: 1002
      responses:
        '200':
          description: 评论成功，返回评论ID
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: integer
                        format: int64
                        description: 新评论的 commentId
              example:
                code: 200
                message: 操作成功
                data: 3001
  /api/content/comments/{commentId}:
    delete:
      tags: [评论]
      summary: 删除评论
      description: |
        删除指定评论。
        - 评论作者可删除自己的评论。
        - 管理员（role >= 2）可删除任意评论。
      operationId: deleteComment
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          description: 评论ID
          schema:
            type: integer
            format: int64
            example: 3001
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/content/comments/{commentId}/like:
    post:
      tags: [评论]
      summary: 点赞评论
      description: 对指定评论点赞。需要登录。
      operationId: likeComment
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          description: 评论ID
          schema:
            type: integer
            format: int64
            example: 3001
      responses:
        '200':
          description: 点赞成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  # ==================== 文件模块 ====================
...
  /api/agent/detect/trigger:
    post:
      tags: [Agent检测]
      summary: 手动触发水军检测
      description: |
        通知 Python Agent 服务立即对当前平台的评论数据进行水军群组分析。
        - **需要管理员权限（role >= 2）**
        - 任务异步执行，接口立即返回；检测结果将自动写入禁言记录
        - 内部通过 RabbitMQ 发送检测指令，若消息队列不可用则降级为 HTTP 直调
      operationId: triggerDetect
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 检测任务已提交
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: string
                        description: 提交结果描述
              example:
                code: 200
                message: 操作成功
                data: "检测任务已提交"
        '403':
          description: 无管理员权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              example:
                code: 403
                message: 无权限访问
                data: null
  /api/agent/mute/records:
    get:
      tags: [Agent检测]
      summary: 查询禁言记录列表（分页）
      description: |
        分页查询所有禁言记录，按创建时间倒序排列。
        **需要版主及以上权限（role >= 1）。**
      operationId: getMuteRecords
      security:
        - BearerAuth: []
      parameters:
        - name: pageNum
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            default: 1
            example: 1
        - name: pageSize
          in: query
          required: false
          description: 每页数量
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageInfo'
                          - type: object
                            properties:
                              records:
                                type: array
                                items:
                                  $ref: '#/components/schemas/AgentMuteRecord'
              example:
                code: 200
                message: 操作成功
                data:
                  records:
                    - muteId: 4001
                      mutedUserId: 1005
                      muteReason: "系统检测到该用户属于水军群组，协同刷评论"
                      muteDays: 7
                      muteType: 0
                      status: 1
                      createdTime: "2024-06-15T16:00:00"
                  total: 42
                  size: 10
                  current: 1
  /api/agent/mute:
    post:
      tags: [Agent检测]
      summary: 手动执行禁言
      description: |
        管理员手动对指定用户执行禁言操作，同时创建禁言记录。
        **需要版主及以上权限（role >= 1）。**
      operationId: manualMute
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          description: 要禁言的用户ID
          schema:
            type: integer
            format: int64
            example: 1005
        - name: reason
          in: query
          required: true
          description: 禁言原因
          schema:
            type: string
            example: "多次发布虚假营销评论"
        - name: days
          in: query
          required: false
          description: 禁言天数（默认7天）
          schema:
            type: integer
            default: 7
            example: 30
      responses:
        '200':
          description: 禁言成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 200
                message: 操作成功
                data: null
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
  /api/agent/mute/{muteId}:
    delete:
      tags: [Agent检测]
      summary: 解除禁言
      description: |
        解除指定禁言记录，将禁言状态改为已解除，并恢复用户正常状态。
        **需要版主及以上权限（role >= 1）。**
      operationId: unmute
      security:
        - BearerAuth: []
      parameters:
        - name: muteId
          in: path
          required: true
          description: 禁言记录ID
          schema:
            type: integer
            format: int64
            example: 4001
      responses:
        '200':
          description: 解除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 200
                message: 操作成功
                data: null
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'